implement feedback adjustments
- [CORE-599]
